requirements.txt
beautifulsoup4~=4.10.0
requests~=2.26.0
pytest~=6.2.5
setuptools~=57.0.0
bs4~=0.0.1
lxml~=4.6.3
python-dateutil~=2.8.2
build~=0.7.0
Jinja2~=3.0.2
fpdf~=1.7.2
colorama~=0.4.4